+++ src/core/app.tsx
@@ -2,35 +2,49 @@ import { __ } from '@wordpress/i18n';
 import { search } from '@wordpress/icons';
 import { dispatch, select } from '@wordpress/data';
 import { useState, useEffect } from '@wordpress/element';
-import { Modal, TextControl, ToggleControl, Button } from '@wordpress/components';
+import { Modal, TextControl, ToggleControl, Button, Tooltip } from '@wordpress/components';
 
-import './styles/app.scss';
+import '../styles/app.scss';
 
-import { getAllowedBlocks, isCaseSensitive } from './utils';
+import { getAllowedBlocks, getBlockEditorIframe, isCaseSensitive, inContainer } from './utils';
 import { Shortcut } from './shortcut';
 
 /**
  * Search & Replace for Block Editor.
  *
  * This function returns a JSX component that comprises
- * the WP Main dashboard, FullscreenModeClose, Modal & Search button.
+ * the Tooltip, Search Icon, Modal & Shortcut.
  *
  * @since 1.0.0
  *
  * @returns {JSX.Element}
  */
-const SearchReplaceForBlockEditor = () => {
+const SearchReplaceForBlockEditor = (): JSX.Element => {
   const [replacements, setReplacements] = useState(0);
   const [isModalVisible, setIsModalVisible] = useState(false);
   const [searchInput, setSearchInput] = useState('');
   const [replaceInput, setReplaceInput] = useState('');
   const [caseSensitive, setCaseSensitive] = useState(false);
 
+  /**
+   * Open Modal.
+   *
+   * @since 1.0.0
+   *
+   * @returns {void}
+   */
   const openModal = (): void => {
     setIsModalVisible(true);
     setReplacements(0);
   }
 
+  /**
+   * Close Modal.
+   *
+   * @since 1.0.0
+   *
+   * @returns {void}
+   */
   const closeModal = (): void => {
     setIsModalVisible(false);
     setReplacements(0);
@@ -39,66 +53,21 @@ const SearchReplaceForBlockEditor = () => {
   /**
    * On Selection.
    *
-   * Populate the find field when the user selects
+   * Populate the search field when the user selects
    * a text range in the Block Editor.
    *
    * @since 1.2.0
    *
    * @returns {void}
    */
-  const onSelection = () => {
-    const selectedText = window.getSelection().toString();
-    const modalSelector = '.search-replace-modal';
+  const handleSelection = (): void => {
+    const selectedText: string = getBlockEditorIframe().getSelection().toString();
+    const modalSelector: string = '.search-replace-modal';
 
     if (selectedText && !inContainer(modalSelector)) {
       setSearchInput(selectedText);
     }
   };
-
-  /**
-   * Check if the selection is made inside target container,
-   * for e.g. the `search-replace-modal`.
-   * 
-   * @since 1.2.1
-   * 
-   * @param {string} selector Target selector.
-   * 
-   * @returns {boolean}
-   */
-  const inContainer = (selector) => {
-    const selection = window.getSelection();
-    const targetDiv = document.querySelector(selector);
-    
-    if (!selection.rangeCount || !targetDiv) {
-      return false;
-    }
-    
-    const range = selection.getRangeAt(0);
-
-    return targetDiv.contains(range.startContainer) && targetDiv.contains(range.endContainer);
-  }
-
-  /**
-   * Listen for Selection.
-   *
-   * Constantly listen for when the user selects a
-   * a text in the Block Editor.
-   *
-   * @since 1.2.0
-   *
-   * @returns {void}
-   */
-  useEffect(() => {
-    document.addEventListener(
-      'selectionchange', onSelection
-    );
-
-    return () => {
-      document.removeEventListener(
-        'selectionchange', onSelection
-      );
-    };
-  }, []);
 
   /**
    * Handle case sensitive toggle feature
@@ -108,6 +77,7 @@ const SearchReplaceForBlockEditor = () => {
    * @since 1.1.0
    *
    * @param {boolean} newValue
+   * @returns {void}
    */
   const handleCaseSensitive = (newValue: boolean): void => {
     setCaseSensitive( newValue );
@@ -128,7 +98,7 @@ const SearchReplaceForBlockEditor = () => {
       return;
     }
 
-    const pattern = new RegExp(
+    const pattern: RegExp = new RegExp(
       `(?<!<[^>]*)${searchInput}(?<![^>]*<)`,
       isCaseSensitive() || caseSensitive ? 'g' : 'gi'
     );
@@ -147,12 +117,12 @@ const SearchReplaceForBlockEditor = () => {
    * @since 1.0.1 Handle edge-cases for quote, pullquote & details block.
    *
    * @param {Object} element Gutenberg editor block.
-   * @param {string} pattern Search pattern.
+   * @param {RegExp} pattern Search pattern.
    * @param {string} text    Replace pattern.
    *
    * @returns {void}
    */
-  const recursivelyReplace = (element, pattern, text) => {
+  const recursivelyReplace = (element, pattern, text): void => {
     if (getAllowedBlocks().indexOf(element.name) !== -1) {
       const args = { element, pattern, text };
 
@@ -191,7 +161,7 @@ const SearchReplaceForBlockEditor = () => {
    *
    * @returns {void}
    */
-  const replaceBlockAttribute = (args, attribute) => {
+  const replaceBlockAttribute = (args, attribute): void => {
     const { attributes, clientId } = args.element;
 
     if (undefined === attributes || undefined === attributes[attribute]) {
@@ -221,14 +191,40 @@ const SearchReplaceForBlockEditor = () => {
     }
   }
 
+  /**
+   * Listen for Selection.
+   *
+   * Constantly listen for when the user selects a
+   * a text in the Block Editor.
+   *
+   * @since 1.2.0
+   *
+   * @returns {void}
+   */
+  useEffect(() => {
+    const editor = getBlockEditorIframe();
+
+    editor.addEventListener(
+      'selectionchange', handleSelection
+    );
+
+    return () => {
+      editor.removeEventListener(
+        'selectionchange', handleSelection
+      );
+    };
+  }, []);
+
   return (
     <>
       <Shortcut onKeyDown={openModal} />
-      <Button
-        icon={ search }
-        label={__('Search & Replace', 'search-replace-for-block-editor')}
-        onClick={openModal}
-      />
+      <Tooltip text={__('Search & Replace', 'search-replace-for-block-editor')}>
+        <Button
+          icon={ search }
+          label={__('Search & Replace', 'search-replace-for-block-editor')}
+          onClick={openModal}
+        />
+      </Tooltip>
       {
         isModalVisible && (
           <Modal