+++ src/core/utils.tsx
@@ -114,23 +114,27 @@ export const isCaseSensitive = (): boolean => {
  *
  * @since 1.2.0
  *
- * @returns Promise<HTMLElement>
+ * @returns Promise<HTMLElement | Error>
  */
-export const getEditorRoot = () => {
-  let elapsedTime = 0;
-  const interval = 100;
+export const getEditorRoot = (): Promise<HTMLElement | Error> => {
+  let elapsedTime: number = 0;
+  const interval: number = 100;
+
+  const selector: string = isWpVersion('6.6.0')
+    ? '.editor-header__toolbar'
+    : '.edit-post-header__toolbar';
 
   return new Promise((resolve, reject) => {
     const intervalId = setInterval(() => {
       elapsedTime += interval;
-      const root = document.getElementById('editor').querySelector('.editor-header__toolbar');
+      const root = document.querySelector(selector) as HTMLElement | null;
 
       if (root) {
         clearInterval(intervalId);
         resolve(root);
       }
 
-      if (elapsedTime > (10 * interval)) {
+      if (elapsedTime > (600 * interval)) {
         clearInterval(intervalId);
         reject(new Error('Unable to get Editor root container...'));
       }
@@ -149,10 +153,90 @@ export const getEditorRoot = () => {
  * @param {HTMLElement} parent - The Parent DOM element.
  * @returns {HTMLDivElement}
  */
-export const getAppRoot = (parent) => {
-  const container = document.createElement('div');
+export const getAppRoot = (parent: HTMLElement): HTMLDivElement => {
+  const container: HTMLDivElement = document.createElement('div');
   container.id = 'search-replace';
   parent.appendChild(container);
 
   return container;
 };
+
+/**
+ * Get iFrame Document.
+ *
+ * Retrieves the document object of the Block Editor
+ * iframe with the name "editor-canvas".
+ *
+ * @since 1.2.1
+ *
+ * @returns {Document}
+ */
+export const getBlockEditorIframe = (): Document => {
+  const editor = document.querySelector('iframe[name="editor-canvas"]');
+
+  return editor && editor instanceof HTMLIFrameElement
+    ? editor.contentDocument || editor.contentWindow?.document
+    : document;
+}
+
+/**
+ * Check if the selection is made inside a Container,
+ * for e.g. the `search-replace-modal`.
+ *
+ * @since 1.2.1
+ *
+ * @param {string} selector Target selector.
+ * @returns {boolean}
+ */
+export const inContainer = (selector: string): boolean => {
+  const selection = window.getSelection() as Selection | null;
+  const targetDiv = document.querySelector(selector) as HTMLElement | null;
+
+  if (!selection?.rangeCount || !targetDiv) {
+    return false;
+  }
+
+  const range: Range = selection.getRangeAt(0);
+
+  return targetDiv.contains(range.startContainer) && targetDiv.contains(range.endContainer);
+}
+
+/**
+ * Check if it's up to WP version.
+ *
+ * @since 1.2.2
+ *
+ * @param {string} version WP Version.
+ * @returns {boolean}
+ */
+export const isWpVersion = (version: string) => {
+  const { wpVersion } = srfbe as { wpVersion: string };
+
+  const argVersion: number = getNumberToBase10(
+    version.split('.').map(Number)
+  );
+
+  const sysVersion: number = getNumberToBase10(
+    wpVersion.split('.').map(Number)
+  );
+
+  return ! (sysVersion < argVersion);
+}
+
+/**
+ * Given an array of numbers, get the Radix
+ * (converted to base 10). For e.g. [5, 6, 1] becomes
+ * 561 or [2, 7, 4] becomes 274.
+ *
+ * @since 1.2.2
+ *
+ * @param {number[]} values Array of positive numbers.
+ * @returns {number}
+ */
+export const getNumberToBase10 = (values: number[]): number => {
+  const radix: number = values.reduce((sum: number, value: number, index: number) => {
+    return sum + (value * Math.pow(10, ((values.length - 1) - index)));
+  }, 0);
+
+  return radix;
+}