+++ src/index.tsx
@@ -1,13 +1,17 @@
+import ReactDOM from 'react-dom';
 import { createRoot } from 'react-dom/client';
 
-import SearchReplaceForBlockEditor from './app';
-import { getAppRoot, getEditorRoot } from './utils'
+import SearchReplaceForBlockEditor from './core/app';
+import { getAppRoot, getEditorRoot, isWpVersion } from './core/utils'
 
 (async () => {
   try {
-    const container = await getEditorRoot();
-    const app = getAppRoot(container);
-    createRoot(app).render(<SearchReplaceForBlockEditor />);
+    const app = getAppRoot(await getEditorRoot() as HTMLElement);
+    if (!isWpVersion('6.2.0')) {
+      ReactDOM.render(<SearchReplaceForBlockEditor />, app);
+    } else {
+      createRoot(app).render(<SearchReplaceForBlockEditor />);
+    }
   } catch (e) {
     console.error(e);
   }