+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 # Changelog
 
+## 1.2.2
+* Fix style issues for WP 6.6.2.
+* Fix timeout issues causing Icon not to load.
+* Fix backward compatibility issues due to WP upgrade.
+* Apply coding standards.
+* Update README notes.
+* Tested up to WP 6.7.1.
+
 ## 1.2.1
 * Fix in modal selection issue.
-* Tested up to WP 6.7.0.
+* Fix missing tooltip component.
+* Fix block editor selection issue due to iframe.
+* Update README notes.
+* Tested up to WP 6.7.1.
 
 ## 1.2.0
 * Fix WP upgrade 6.7 issues.